% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{aggregate_results_using_Rubins_rules}
\alias{aggregate_results_using_Rubins_rules}
\title{Combine results using Rubin's rules}
\usage{
aggregate_results_using_Rubins_rules(
  MAR_model,
  M,
  complete_datasets,
  idx_of_coef_of_interest,
  null_value = 0,
  alpha = 0.05
)
}
\arguments{
\item{alpha}{}
}
\description{
Combine results using Rubin's rules
}
