% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_imputed_data.R
\name{calculate_epsilon}
\alias{calculate_epsilon}
\title{Calculate epsilon}
\usage{
calculate_epsilon(N, trace_of_hat_matrix, sigma_hat_squared)
}
\arguments{
\item{sigma_hat_squared}{}
}
\description{
Calculate epsilon
}
