% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{calculate_theta_and_variance_hats}
\alias{calculate_theta_and_variance_hats}
\title{Title}
\usage{
calculate_theta_and_variance_hats(
  cal,
  M,
  complete_datasets,
  idx_of_coef_of_interest,
  call
)
}
\arguments{
\item{call}{}
}
\description{
Title
}
