% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_imputed_data.R
\name{generate_complete_dataset_using_multiple_imputation}
\alias{generate_complete_dataset_using_multiple_imputation}
\title{Generate complete imputed dataset}
\usage{
generate_complete_dataset_using_multiple_imputation(
  MAR_model,
  design_X_full,
  k,
  uids,
  time_var,
  outcome_var,
  timepoints,
  Beta_m,
  beta_i_ms
)
}
\arguments{
\item{beta_i_ms}{}
}
\description{
Generate complete imputed dataset
}
