% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{generate_random_Beta_m}
\alias{generate_random_Beta_m}
\title{Draw random Beta M, for fixed effects}
\usage{
generate_random_Beta_m(uids, mod)
}
\arguments{
\item{uids}{A vector of uids in the provided dataset}

\item{mod}{An instance of class lmerMod from the package lme4 (fit using \code{lmer})}
}
\value{
A random draw of Beta_m for use in multiple imputation
}
\description{
Draw random Beta M, for fixed effects
}
