% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_inferential_statistics_from_lmm}
\alias{extract_inferential_statistics_from_lmm}
\title{Title}
\usage{
extract_inferential_statistics_from_lmm(mod, idx_of_coef_of_interest)
}
\arguments{
\item{idx_of_coef_of_interest}{}
}
\description{
Title
}
