% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_design_X_full}
\alias{generate_design_X_full}
\title{Title}
\usage{
generate_design_X_full(
  MAR_model,
  uids,
  timepoints,
  time_var,
  outcome_var,
  type = "matrix"
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
